refactor azure profile data on config store attr
