+++ src/client/src/store/azureProfileData/azure/selector.ts
@@ -1,7 +1,7 @@
 import { createSelector } from "reselect";
 import { AppState } from "../../combineReducers";
 
-const getSubscriptionsSelector = (state: AppState): Subscription[] => state.azureProfileData.profileData.subscriptions;
+const getSubscriptionsSelector = (state: AppState): Subscription[] => state.config.azureProfileData.subscriptions;
 
 const getDropdownSubscriptionsSelector = createSelector([getSubscriptionsSelector], subscriptions => {
   return subscriptions.map(subscription => {
