+++ src/client/src/store/azureProfileData/combineReducers.ts
@@ -1,9 +1,8 @@
 import { combineReducers } from "redux";
-import {profileData, isLoggedIn} from "./login/reducer";
+import { isLoggedIn } from "./login/reducer";
 import { subscriptionData } from "./azure/reducer";
 
 export default combineReducers({
   isLoggedIn,
-  profileData,
   subscriptionData
 });
