+++ src/client/src/store/azureProfileData/typeKeys.ts
@@ -1,6 +1,5 @@
 export enum AZURE_TYPEKEYS {
   IS_LOGGED_IN_TO_AZURE = "WTS/azure/IS_LOGGED_IN_TO_AZURE",
-  LOG_IN_TO_AZURE = "WTS/azure/LOG_IN_TO_AZURE",
   LOG_OUT_OF_AZURE = "WTS/azure/LOG_OUT_OF_AZURE",
   SAVE_COSMOS_DB_RESOURCE_SETTINGS = "WTS/azure/SAVE_COSMOS_DB_RESOURCE_SETTINGS",
   SAVE_APP_SERVICE_SETTINGS = "WTS/azure/SAVE_APP_SERVICE_SETTINGS",
