+++ src/client/src/store/config/azure/model.ts
@@ -1,5 +1,5 @@
-import { AZURE_TYPEKEYS } from "../typeKeys";
 import { CONFIG_TYPEKEYS } from "../../typeKeys";
+import { AZURE_TYPEKEYS } from "../../azureProfileData/typeKeys";
 
 export interface IsLoggedIntoAzureAction {
   type: AZURE_TYPEKEYS.IS_LOGGED_IN_TO_AZURE;
