+++ src/client/src/store/config/combineReducers.ts
@@ -4,13 +4,15 @@ import previewStatus from "./config/previewReducer";
 import validations from "./validations/setValidationsReducer";
 import isValidatingName from "./validations/validatingNameReducer";
 import versions from "./versions/reducer";
+import { azureProfileData } from "./azure/reducer";
 
 const configReducer = combineReducers({
   detailsPage,
   previewStatus,
   validations,
   isValidatingName,
-  versions
+  versions,
+  azureProfileData
 });
 
 export default configReducer;
