+++ src/client/src/store/typeKeys.ts
@@ -22,4 +22,5 @@ export enum TEMPLATES_TYPEKEYS {
 export enum CONFIG_TYPEKEYS {
   SET_PREVIEW_STATUS = "WTS/config/SET_PREVIEW_STATUS",
   RESET_WIZARD = "WTS/config/RESET_WIZARD",
+  LOG_IN_TO_AZURE = "WTS/azure/LOG_IN_TO_AZURE"
 }
